--- a/codacy-analysis-cli
+++ b/codacy-analysis-cli
@@ -1,0 +1,1 @@
+Not Found