## Summary

Implements fixed 15s/5s round timer warning thresholds as requested in Issue #241. Replaces the previous percentage-based warning system with consistent fixed time thresholds that work reliably across different round time limits.

## Changes Made

### ✅ Core Implementation
- **Added constants**: `WARNING_THRESHOLD = 15000ms` and `CRITICAL_THRESHOLD = 5000ms`
- **Updated warning logic**: Replaced percentage-based thresholds with fixed time thresholds
- **Simplified calculation**: `calculateThresholdStates` now uses simple boolean logic with fixed values

### ✅ Testing
- **Added comprehensive tests**: 6 new test cases for fixed threshold behavior
- **Verified edge cases**: Tests for exactly 15s, 5s, and various time remaining scenarios
- **Cross-validated**: Tests work with different round time limits (30s, 60s, 120s)

### ✅ Visual Behavior
- **Warning state**: Shows amber/yellow when ≤15 seconds remaining
- **Critical state**: Shows red when ≤5 seconds remaining  
- **Expired state**: Shows expired styling when time reaches 0
- **Consistent UX**: Works reliably regardless of total round time configured

## Technical Details

**Before**: Percentage-based thresholds (25% warning, 10% critical)
```typescript
const warningThreshold = roundTimeLimit * 0.25;
const criticalThreshold = roundTimeLimit * 0.1;
```

**After**: Fixed time thresholds
```typescript
const WARNING_THRESHOLD = 15000; // 15 seconds
const CRITICAL_THRESHOLD = 5000;  // 5 seconds
isRoundWarning: remaining <= WARNING_THRESHOLD && remaining > CRITICAL_THRESHOLD
```

## Testing Strategy

### Unit Tests
- ✅ All 32 existing tests continue to pass
- ✅ 6 new tests specifically for fixed threshold behavior
- ✅ Edge case testing (exactly 15s, 5s, 0s)
- ✅ Multiple round time limit validation

### Quality Gates
- ✅ ESLint: No warnings or errors
- ✅ TypeScript: Compiles without errors
- ✅ Build: Production build succeeds
- ✅ Tests: 100% test pass rate
- ✅ Code Quality: Meets all coding standards

## User Experience Impact

### DM Benefits
- **Predictable warnings**: Always get warning at 15s regardless of round length
- **Consistent timing**: Critical alert always at 5s for urgent action
- **Better planning**: Fixed thresholds help DMs manage encounter pacing

### Visual Indicators
- **Normal**: Default timer styling
- **Warning** (≤15s): Amber/yellow indicator
- **Critical** (≤5s): Red indicator
- **Expired** (0s): Expired state styling

## Compliance

### Code Quality
- ✅ No increase in cyclomatic complexity
- ✅ Functions under 50 lines
- ✅ File under 450 lines
- ✅ Descriptive variable names
- ✅ No code duplication
- ✅ Comprehensive test coverage

### Requirements Satisfied
- [x] Round timer shows warning state (amber) when ≤15 seconds remaining
- [x] Round timer shows critical state (red) when ≤5 seconds remaining  
- [x] Round timer shows expired state when time reaches 0
- [x] Warnings only activate when roundTimeLimit is configured
- [x] No increase in cyclomatic complexity
- [x] Simple boolean calculations only
- [x] All existing tests continue to pass
- [x] New tests added for warning state calculations

CLOSES #241

🤖 Generated with [Claude Code](https://claude.ai/code)